more rearranging, fixing what was arranged
