+++ persim/__init__.py
@@ -5,12 +5,6 @@ from .wasserstein import *
 from .heat import *
 from .gromov_hausdorff import *
 from .visuals import *
-from .landscapes import (
-    PersLandscapeExact, 
-    PersLandscapeApprox, 
-    PersistenceLandscaper,  
-    plot_landscape, 
-    plot_landscape_simple
-)
+from .landscapes import *
 
 from ._version import __version__
