+++ persim/landscapes/__init__.py
@@ -2,3 +2,4 @@ from .exact import PersLandscapeExact
 from .approximate import PersLandscapeApprox
 from .transformer import PersistenceLandscaper
 from .visuals import plot_landscape, plot_landscape_simple
+from .tools import *
