+++ persim/landscapes/auxiliary.py
@@ -5,7 +5,11 @@
 import itertools
 import numpy as np
 
-__all__ = ["death_vector", "linear_combination"]
+
+__all__ = [
+    "death_vector",
+    "linear_combination",
+]
 
 
 def death_vector(dgms: list, hom_deg: int = 0):
