+++ persim/landscapes/transformer.py
@@ -5,8 +5,9 @@
 from operator import itemgetter
 import numpy as np
 from sklearn.base import BaseEstimator, TransformerMixin
-from .exact import PersLandscapeExact
+
 from .approximate import PersLandscapeApprox
+from .exact import PersLandscapeExact
 
 
 __all__ = ["PersistenceLandscaper"]
@@ -28,7 +29,7 @@ class PersistenceLandscaper(BaseEstimator, TransformerMixin):
 
     num_steps : int, optional
         Number of steps of approximating grid.
-    
+
     flatten : bool, optional
         Determines if the resulting values are flattened.
 
@@ -82,11 +83,11 @@ class PersistenceLandscaper(BaseEstimator, TransformerMixin):
             return f"PersistenceLandscaper(hom_deg={self.hom_deg}, start={self.start}, stop={self.stop}, num_steps={self.num_steps})"
 
     def fit(self, dgms):
-        """ Find optimal `start` and `stop` parameters for approximating grid. 
-        
+        """Find optimal `start` and `stop` parameters for approximating grid.
+
         Parameters
         ----------
-        
+
         dgms : list of (-,2) numpy.ndarrays
             List of persistence diagrams
         """
@@ -99,20 +100,20 @@ class PersistenceLandscaper(BaseEstimator, TransformerMixin):
         return self
 
     def transform(self, dgms):
-        """ Construct persistence landscape values.
-        
+        """Construct persistence landscape values.
+
         Parameters
         ----------
-        
+
         dgms : list of (-,2) numpy.ndarrays
             List of persistence diagrams
-        
+
         flatten : bool, optional
             Flag determining whether output values are flattened
-        
+
         Returns
         -------
-        
+
         numpy.ndarray
             Persistence Landscape values sampled on approximating grid.
         """
